refactor: implement TaskGroupLevel pattern to replace isTopLevel
- Introduced TaskGroupLevel enum with ROOT and NESTED configurations
- Each level has formatTime, showTitle, headerOrder, and cssClass properties
- Replaced boolean isTopLevel checks with level-based design pattern
- More extensible design that supports additional hierarchy levels
- Fixed null check for filesToggle element in setupEventListeners

This provides a cleaner, more object-oriented approach to handling
different task group hierarchy levels with their associated behaviors.
